requirements.txt
chromadb
lightgbm
shap
langchain
pandas
scikit-learn
sentence-transformers
python-dotenv
fastapi
uvicorn[standard]
python-multipart